bot-keydrop-main/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
discord-webhook==1.3.0
pywin32>=307
pyinstaller==6.3.0
psutil==5.9.6
pillow==10.1.0
cryptography>=41.0.8
colorama==0.4.6
packaging>=23.2
setuptools>=69.0.3
wheel>=0.42.0
customtkinter==5.2.2
